[#35] Moved `NamedTuple` into `complex.py`.
It can contain other filters now, so it fits.
