+++ test/complex_test.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import, division, print_function, \
     unicode_literals
 
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 
 import filters as f
 from filters.test import BaseFilterTestCase
@@ -1040,5 +1040,158 @@ class FilterMapperTestCase(BaseFilterTestCase):
                     'type': None,
                     'data': None,
                 },
+            },
+        )
+
+
+Color = namedtuple('Color', ('r', 'g', 'b'))
+
+
+class NamedTupleTestCase(BaseFilterTestCase):
+    @staticmethod
+    def filter_type():
+        return f.NamedTuple(Color)
+
+    def test_success_none(self):
+        """
+        ``None`` always passes this filter.
+
+        Chain with :py:class:`f.Required` if you want to disallow null
+        values.
+        """
+        self.assertFilterPasses(None)
+
+    def test_success_namedtuple_correct_type(self):
+        """
+        Incoming value is already a namedtuple of the expected type.
+        """
+        self.assertFilterPasses(Color(64, 128, 192))
+
+    def test_success_namedtuple_different_type(self):
+        """
+        Incoming value is a namedtuple instance, but of a different
+        type.
+
+        Since namedtuples are still tuples, this has the same result as
+        for any other incoming iterable.
+        """
+        # Just to be tricky, we'll make it look very close to the
+        # expected type.
+        AltColor = namedtuple('AltColor', Color._fields)
+
+        self.assertFilterPasses(
+            AltColor(64, 128, 192),
+            Color(64, 128, 192),
+        )
+
+    def test_success_iterable(self):
+        """
+        Incoming value is an iterable with correct values.
+        """
+        value = [64, 128, 192]
+        self.assertFilterPasses(value, Color(*value))
+
+    def test_success_iterable_compat(self):
+        """
+        Incoming value is an iterable, formatted for compatibility with
+        Python < 3.6.
+        """
+        value = [('b', 192), ('g', 128), ('r', 64)]
+        self.assertFilterPasses(value, Color(*value))
+
+    def test_success_mapping(self):
+        """
+        Incoming value is a mapping with correct keys.
+        """
+        value = {'r': 64, 'g': 128, 'b': 192}
+        self.assertFilterPasses(value, Color(**value))
+
+    def test_fail_incompatible_type(self):
+        """
+        Incoming value has a type that we cannot work with.
+        """
+        self.assertFilterErrors(42, [f.Type.CODE_WRONG_TYPE])
+
+    def test_fail_iterable_too_short(self):
+        """
+        Incoming value is an iterable that is missing one or more
+        values.
+        """
+        self.assertFilterErrors((64, 128,), [f.MinLength.CODE_TOO_SHORT])
+
+    def test_fail_iterable_too_long(self):
+        """
+        Incoming value is an iterable that has too many values.
+        """
+        self.assertFilterErrors(
+            (64, 128, 192, 0.5),
+            [f.MaxLength.CODE_TOO_LONG],
+        )
+
+    def test_fail_mapping_missing_keys(self):
+        """
+        Incoming value is a mapping that is missing one or more keys.
+        """
+        self.assertFilterErrors(
+            {},
+
+            {
+                'r': [f.FilterMapper.CODE_MISSING_KEY],
+                'g': [f.FilterMapper.CODE_MISSING_KEY],
+                'b': [f.FilterMapper.CODE_MISSING_KEY],
+            },
+        )
+
+    def test_fail_mapping_extra_keys(self):
+        """
+        Incoming value is a mapping that has extra keys that we don't
+        know what to do with.
+        """
+        self.assertFilterErrors(
+            {
+                'r': 64,
+                'g': 128,
+                'b': 192,
+                'a': 0.5,
+            },
+
+            {
+                'a': [f.FilterMapper.CODE_EXTRA_KEY],
+            },
+        )
+
+    def test_success_filter_map(self):
+        """
+        Applying a :py:class:`f.FilterMap` to the values in a namedtuple
+        after converting (success case).
+        """
+        self.filter_type = lambda: f.NamedTuple(Color, {
+            # For whatever reason, we decide not to filter ``r``.
+            'g': f.Required | f.Int | f.Min(0) | f.Max(255),
+            'b': f.Required | f.Int | f.Min(0) | f.Max(255),
+        })
+
+        self.assertFilterPasses(
+            ('64.0', '128', 192.0),
+            Color('64.0', 128, 192),
+        )
+
+    def test_fail_filter_map(self):
+        """
+        Applying a :py:class:`f.FilterMap` to the values in a namedtuple
+        after converting (failure case).
+        """
+        self.filter_type = lambda: f.NamedTuple(Color, {
+            # For whatever reason, we decide not to filter ``r``.
+            'g': f.Required | f.Int | f.Min(0) | f.Max(255),
+            'b': f.Required | f.Int | f.Min(0) | f.Max(255),
+        })
+
+        self.assertFilterErrors(
+            ['NaN', None, (42,)],
+
+            {
+                'g': [f.Required.CODE_EMPTY],
+                'b': [f.Decimal.CODE_INVALID],
             },
         )
